requirements.txt
sentence-transformers
pinecone-client
pandas
FastAPI
uvicorn